# Replace LinkedList with array, remove LinkedList