… packages/core/src/LinkedList.js