## Summary

LinkedList can be deleted with no significant penalty.

## Details

LinkedList is a doubly-linked list that allowed constant time removal from the middle.  It is only used in `mergeConcurrently` to manage the fact that inner streams may end at any time, and thus need to be removed from the `current` set of active inner streams.  In previous versions of node/v8, LinkedList's constant time removal made an observable difference in the performance of `mergeConcurrently` (and thus `chain` and `concatMap`) at scale (large numbers of active inner streams).

On the downside, LinkedList is more code to maintain, and imposed the extra requirement onto `InnerSink` that it have the shape of a LinkedList node, specifically `prev` and `next` fields.

On node 8 and 10, a linear array scan seems to be approaching LinkedList, at least for the orders of magnitude in the perf tests (1000 inner streams).  See below for representative results on node 10.

Deleting LinkedList with no significant penalty seems like a win, but want to get everyone's take on it.

## Perf results

### 1k inner streams

Array:

```
chain 1000 x 1000 streams
-------------------------------------------------------
most                 84.58 op/s ±  4.30%   (76 samples)

concatMap 1000 x 1000 streams
-------------------------------------------------------
most                 77.82 op/s ±  5.26%   (71 samples)
```

LinkedList:

```
chain 1000 x 1000 streams
-------------------------------------------------------
most                 86.45 op/s ±  3.73%   (77 samples)

concatMap 1000 x 1000 streams
-------------------------------------------------------
most                 78.16 op/s ±  4.85%   (71 samples)
```

### 100 inner streams

Curiously, array is consistently faster for concatMap in this scenario:

Array:

```
chain 100 x 1000 streams
-------------------------------------------------------
most                566.28 op/s ± 18.52%   (82 samples)

concatMap 100 x 1000 streams
-------------------------------------------------------
most                421.13 op/s ± 12.97%   (82 samples)
```

LinkedList:

```
chain 100 x 1000 streams
-------------------------------------------------------
most                571.84 op/s ± 22.17%   (80 samples)

concatMap 100 x 1000 streams
-------------------------------------------------------
most                370.77 op/s ±  2.90%   (73 samples)
```

### 10k inner streams

Array:

```

chain 10000 x 1000 streams
-------------------------------------------------------
most                  9.13 op/s ±  3.43%   (46 samples)

concatMap 10000 x 1000 streams
-------------------------------------------------------
most                  8.64 op/s ±  1.72%   (44 samples)
```

LinkedList:

```
chain 10000 x 1000 streams
-------------------------------------------------------
most                  9.24 op/s ±  1.97%   (46 samples)

concatMap 10000 x 1000 streams
-------------------------------------------------------
most                  8.74 op/s ±  2.53%   (44 samples)
```

## Size comparison

Here are the built size differences in bytes:  About 1.5k unminified, and about 200 bytes min+gzip:

After deleting LinkedList:

```
❯ ls -Bl dist
.rw-r--r--  82,502 brian 13 Aug  8:01 index.es.js
.rw-r--r-- 123,876 brian 13 Aug  8:01 index.es.js.map
.rw-r--r--  84,180 brian 13 Aug  8:01 index.js
.rw-r--r--  13,761 brian 11 May 21:22 index.js.flow
.rw-r--r-- 123,976 brian 13 Aug  8:01 index.js.map
.rw-r--r--  10,172 brian 13 Aug  8:01 index.min.js.gz
```

Master (w/LinkedList):

```
❯ ls -Bl dist
.rw-r--r--  83,968 brian 13 Aug  6:54 index.es.js
.rw-r--r-- 125,890 brian 13 Aug  6:54 index.es.js.map
.rw-r--r--  85,635 brian 13 Aug  6:54 index.js
.rw-r--r--  13,761 brian 11 May 21:22 index.js.flow
.rw-r--r-- 125,989 brian 13 Aug  6:54 index.js.map
.rw-r--r--  10,323 brian 13 Aug  6:54 index.min.js.gz
```

UPDATE: Consistently ordered perf results (chain then concatMap), for easier comparison.