@briancavalier 2018-08-13
FYI, I also tried Set and Map.  Array was faster than each at these scales.